Add examples of custom models
Split up example file as it was getting too large.

Resolves #14
